Upd blank_note_task(), note_task(), game_task()
blank_note_task() : miss_penalty logic is added as we need to check if user misses the notes at the end while we pumping blank notes. 
note_task() : For the first five notes, we don't have a correct note as user should not press any button, so the logic is added. Also we don't have a miss penalty for first 5 notes as no note will not reach the end. 
game_task(): Added check press control to LVL_BEG and removed control logic at the LVL_CONT as we handle that at the note_task().